requirements.txt
# Dev environment
pip=25.0.1
autopep8=2.3.2

# App
flask=3.1.0
gunicorn=23.0.0
python-dotenv=1.1.0